Add files via upload
One of the underlying targets of movies is to evoke emotions in their viewers. the scraper is written in Python and uses lxml for parsing the webpages. BeautifulSoup is used for pulling data out of HTML and XML files. There are 8 classes of emotion that would be effective to classify a text. These are: ‘Anger’, ‘Anticipation’, ‘Disgust’, ‘Fear’, ‘Joy’, ‘Sad’, ‘Surprise’, ‘Trust’. these emotions are taken as input and the corresponding movies would be displayed for the emotion. Source site/sites : www.imdb.com